requirements.txt
customtkinter
langchain
llama-cpp-python
pydantic < 2.0
PyYAML
